Correctly generate code with protobuf using CMake
The generated Protobuf code was severely outdated and doesn't belong in
the repo. Only the schema should be under version control as the source
of truth. As part of porting this build to CMake, the Protobuf code
should be generated the correct way using CMake. The outdated
instructions for generating code from the schema was also removed from
the code comments.